// src/app/model/ponto.ts
export interface PontoLista {
  nm_sequenc: number;
  hr_entrada: string;
  hr_partida: string;
}

export interface Registro {
  tipo: string;
  horario: string;
}

export interface PontoVw {
  dt_movimen: string;
  nm_usuario: string;
  hr_entrada1: string;
  hr_partida1: string;
  hr_entrada2: string;
  hr_partida2: string;
  hr_entrada3: string;
  hr_partida3: string;
  hr_entrada4: string;
  hr_partida4: string;
  hr_entrada5: string;
  hr_partida5: string;
}